requirements.txt
ailment==9.0.5327
aiohttp==3.7.3
angr==9.0.5327
archinfo==9.0.5327
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bitstring==3.1.7
cachetools==4.2.0
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
claripy==9.0.5327
cle==9.0.5327
colorama==0.4.4
CppHeaderParser==2.7.4
decorator==4.4.2
discord.py==1.6.0
dpkt==1.9.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
isort==5.7.0
itanium-demangler==1.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
mulpyplexer==0.8
multidict==5.1.0
networkx==2.5
numpy==1.19.5
pefile==2019.4.18
plumbum==1.6.9
ply==3.11
progressbar2==3.53.1
protobuf==3.14.0
psutil==5.8.0
psycopg2==2.8.6
pycparser==2.20
pyelftools==0.27
pylint==2.6.0
PySMT==0.9.0
python-dotenv==0.15.0
python-utils==2.4.0
pyvex==9.0.5327
requests==2.25.1
rpyc==5.0.0
selenium==3.141.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.3.0
soupsieve==2.1
toml==0.10.2
typing-extensions==3.7.4.3
unicorn==1.0.2
urllib3==1.26.2
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
z3-solver==4.8.9.0